% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlock.R
\name{release_lock}
\alias{release_lock}
\title{Release a Redis Lock}
\usage{
release_lock(conn, lockname, identifier)
}
\arguments{
\item{conn}{A connection object to Redis created with the `redux` package.}

\item{lockname}{A string representing the name of the lock to release.}

\item{identifier}{A string representing the unique identifier of the lock.}
}
\value{
A logical value indicating whether the lock was released successfully.
}
\description{
This function attempts to release a lock in Redis using the specified
identifier. It checks if the lock belongs to the identifier before
releasing it.
}
\examples{
\dontrun{
conn <- redux::redis()  # Create a connection to Redis
success <- release_lock(conn, "my_lock_name", lock_id)
}
}
